SRT-28657 Make inner Class static
A change solely to please FindBugs, as we will never serialize the Test
class.
(FindBugs noted: This Serializable class is an inner class of a
non-serializable class. Thus, attempts to serialize it will also attempt
to associate instance of the outer class with which it is associated,
leading to a runtime error.)