SAE-19 Add ActionDispatcher framework
The ActionDispatcher is dealing with two tasks that are done after Cache
mutation: Notification of Listeners and handling Write-Through. The
ActionDispatcher is used in user-initiated actions (JCache API) athat do
Cache mutation and require notification and/or write-through. An
Exception are the bulk operations setAll() and deleteAll() that require
to run write-through BEFORE mutating the Cache.