SAE-150 Fix double releasing and notifcation
If two threads call release() / releaseHolder() in parallel, the Holder will be freed multiple times. This itself is not a problem, as the code is idempotent. But before this change the caller did not know if it was the "first" release of the holder. So mutliple Threads could assume that they removed / released the entry. If it happens to both the  eviction and expiration threads, they would both fire an "EXPIRED" event, thus notifying the caller twice. The fix makes sure that the caller gets informed, whether it is him who actually released the holder.