SAE-164 Fixes on Millisecond precision
The Builder is kept backwards compatible, still using seconds. The Cache
constructor needs to reflect this, which is in this patch.

The Ccahe.getMaxCacheTime() was delivering SECONDS. This is problematic,
as internally we now have milliseconds. To avoid confusion with rounding
or other issues, we will drop the method from the official API. It would
better be placed in the Configuration object if we would want to keep
it.