SAE-164 Eviction notification uses "released" state again
During JCache Eviction implementation the remove() call was changed to
return a slightly different value ("old value
if-valid-before-releasing"). This new remove() semantics are not
appropriate for the eviction thread: It must notify about all elements
it removed, even if they were already expired. Thus implementing a new
removeAndRelease() method for the eviciton thread.