SAE-164 Thread safety for lifecycle events
1) Make CacheManager names unique: uriSeqno must be static, as it is the
sequence number for the CacheManager and not for the caches in the
CacheManager. Otherwise all Cache Manager instances created with the
dafault constructor would be "tcache:/manager-1". The other constructors
are unaffected.

2) Cache.close() removes the Cache also from the associated
CacheManager. This is sensible, as the Cache cannot be used after
close() and also is required by the JCache Spec.

3) Synchronize TCacheFactory.standardFactory()
Concurrent Thread could create two "standard factories". While this
alone is not strictly forbidden, it would be highly unexpected that the
standardFactory changes (unless it would be closed in the menatime).