SAE-95 Set saner cleanUpIntervalMillis defaults
In JSR107 mode (and as a fallback for illegal values), we used 6 seconds
as cleanUpIntervalMillis, which is very low. We use now 1 minute.
Additionally we make the lower limit 1ms. Effectively the
cleanUpIntervalMillis is now in the interval [1ms, 5min]

This commit also contains some cleanups (variable naming, log string)