SAE-95 Fix setExpireUntil()
expireUntil did not work properly, especially if you called it a long
time after the entry was added to the Cache. Offset from now was not
properly added and this is fixed now.
Also short maxDelay values are now taken precisely instead of losing
precision by converting to SECONDS. This is additionally leading to a
better distribution, ass less entries will have identical expiration
times.